Made handle backspace recursive to handle edge cases
